+++ symbol.h
@@ -1,7 +1,7 @@
 /*
  * symbol - global and local symbol routines
  *
- * Copyright (C) 1999  David I. Bell
+ * Copyright (C) 1999-2006  David I. Bell
  *
  * Calc is open software; you can redistribute it and/or modify it under
  * the terms of the version 2.1 of the GNU Lesser General Public License
@@ -17,8 +17,8 @@
  * received a copy with calc; if not, write to Free Software Foundation, Inc.
  * 59 Temple Place, Suite 330, Boston, MA  02111-1307, USA.
  *
- * @(#) $Revision: 29.4 $
- * @(#) $Id: symbol.h,v 29.4 2001/06/08 21:00:58 chongo Exp $
+ * @(#) $Revision: 29.6 $
+ * @(#) $Id: symbol.h,v 29.6 2006/05/19 15:26:10 chongo Exp $
  * @(#) $Source: /usr/local/src/cmd/calc/RCS/symbol.h,v $
  *
  * Under source code control:	1990/02/15 01:48:37
@@ -54,7 +54,7 @@
 
 typedef struct global GLOBAL;
 struct global {
-	int g_len;		/* length of symbol name */
+	size_t g_len;		/* length of symbol name */
 	short g_filescope;	/* file scope level of symbol (0 if global) */
 	short g_funcscope;	/* function scope level of symbol */
 	char *g_name;		/* global symbol name */
